requirements.txt
django>2.2
djangorestframework>3.0.0
selenium
